Dodawnaie i wyswietlanie Albumu
