+++ src/main/java/io/mbab/sda/groupproject2/menu/action/MainAction.java
@@ -1,7 +1,7 @@
-package io.mbab.sda.groupproject.menu.action;
+package io.mbab.sda.groupproject2.menu.action;
 
-import io.mbab.sda.groupproject.menu.CustomScanner;
-import io.mbab.sda.groupproject.menu.MenuActionContext;
+import io.mbab.sda.groupproject2.menu.CustomScanner;
+import io.mbab.sda.groupproject2.menu.MenuActionContext;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -13,11 +13,8 @@ public class MainAction implements MenuAction {
   @Override
   public void execute() {
     System.out.println("0) Zamknij aplikację");
-    System.out.println("1) Dodaj miasto");
-    System.out.println("2) Wyswietl miasta");
-    System.out.println("3) Dodaj piosenkę:");
-    System.out.println("4) Wyswietl piosenki:");
-
+    System.out.println("1) Dodaj Album");
+    System.out.println("2) Wyswietl nazwe Albumu");
 
     var input = scanner.nextLine();
 
@@ -27,7 +24,7 @@ public class MainAction implements MenuAction {
     }
 
     if (input.equals("1")) {
-      ctx.use(CreateCityAction.class).execute();
+      ctx.use(CreateAlbumAction.class).execute();
       return;
     }
 
@@ -35,15 +32,6 @@ public class MainAction implements MenuAction {
       ctx.use(ViewCitiesAction.class).execute();
       return;
     }
-
-    if (input.equals("3")){
-      ctx.use(CreateSongAction.class).execute();
-      return;
-    }
-    if (input.equals("4"))
-      ctx.use(VievSongsAction.class).execute();
-
-
 
     System.out.println("Wprowadzono nieprawidłowa wartość!");
     execute();
